SQMAGOPC-446: add seperate class for commerce edition
